extrapolate file update functions
